+++ src/OpenApiRuntime/Client/Psr7HttplugEndpointTrait.php
@@ -4,26 +4,12 @@ declare(strict_types=1);
 
 namespace Jane\OpenApiRuntime\Client;
 
-use Jane\OpenApiRuntime\Client\Exception\InvalidFetchModeException;
-use Psr\Http\Message\ResponseInterface;
-use Symfony\Component\Serializer\SerializerInterface;
+@trigger_error(sprintf('The "%s" trait is deprecated since Jane 5.1, use "%s" instead.', Psr7HttplugEndpointTrait::class, Psr7EndpointTrait::class), E_USER_DEPRECATED);
 
+/**
+ * @deprecated since Jane 5.1, use Psr7EndpointTrait instead.
+ */
 trait Psr7HttplugEndpointTrait
 {
-    abstract protected function transformResponseBody(string $body, int $status, SerializerInterface $serializer, string $contentType = null);
-
-    public function parsePSR7Response(ResponseInterface $response, SerializerInterface $serializer, string $fetchMode = Client::FETCH_OBJECT)
-    {
-        if ($fetchMode === Client::FETCH_OBJECT) {
-            $contentType = $response->hasHeader('Content-Type') ? current($response->getHeader('Content-Type')) : null;
-
-            return $this->transformResponseBody((string) $response->getBody(), $response->getStatusCode(), $serializer, $contentType);
-        }
-
-        if ($fetchMode === Client::FETCH_RESPONSE) {
-            return $response;
-        }
-
-        throw new InvalidFetchModeException(sprintf('Fetch mode %s is not supported', $fetchMode));
-    }
+    use Psr7EndpointTrait;
 }